Biggest refactor ever
