--- utils.py
@@ -1,116 +1,0 @@
-import numpy as np
-import glob
-import Ska.Sun
-import os
-from Chandra.Time import DateTime
-import Ska.engarchive.fetch_sci as fetch
-import matplotlib.pyplot as plt
-
-def pointpair(x, y=None):
-    if y is None:
-        y = x
-    return np.array([x, y]).reshape(-1, order='F')
-
-def calc_off_nom_rolls(states):
-    off_nom_rolls = []
-    for state in states:
-        att = [state[x] for x in ['q1', 'q2', 'q3', 'q4']]
-        time = (state['tstart'] + state['tstop']) / 2
-        off_nom_rolls.append(Ska.Sun.off_nominal_roll(att, time))
-    return np.array(off_nom_rolls)
-
-def globfile(pathglob):
-    """Return the one file name matching ``pathglob``.  Zero or multiple
-    matches raises an IOError exception."""
-
-    files = glob.glob(pathglob)
-    if len(files) == 0:
-        raise IOError('No files matching %s' % pathglob)
-    elif len(files) > 1:
-        raise IOError('Multiple files matching %s' % pathglob)
-    else:
-        return files[0]
-
-def get_bs_cmds(oflsdir):
-    """Return commands for the backstop file in opt.oflsdir.
-    """
-    import Ska.ParseCM
-    backstop_file = globfile(os.path.join(oflsdir, 'CR*.backstop'))
-    logger.info('Using backstop file %s' % backstop_file)
-    bs_cmds = Ska.ParseCM.read_backstop(backstop_file)
-    logger.info('Found %d backstop commands between %s and %s' %
-                (len(bs_cmds), bs_cmds[0]['date'], bs_cmds[-1]['date']))
-
-    return bs_cmds
-
-def get_telem_values(tstart, msids, days=14, name_map={}):
-    """
-    Fetch last ``days`` of available ``msids`` telemetry values before
-    time ``tstart``.
-
-    :param tstart: start time for telemetry (secs)
-    :param msids: fetch msids list
-    :param days: length of telemetry request before ``tstart``
-    :param dt: sample time (secs)
-    :param name_map: dict mapping msid to recarray col name
-    :returns: np recarray of requested telemetry values from fetch
-    """
-    tstart = DateTime(tstart).secs
-    start = DateTime(tstart - days * 86400).date
-    stop = DateTime(tstart).date
-    logger.info('Fetching telemetry between %s and %s' % (start, stop))
-    msidset = fetch.MSIDset(msids, start, stop, stat='5min')
-    start = max(x.times[0] for x in msidset.values())
-    stop = min(x.times[-1] for x in msidset.values())
-    msidset.interpolate(328.0, start, stop + 1)  # 328 for '5min' stat
-
-    # Finished when we found at least 4 good records (20 mins)
-    if len(msidset.times) < 4:
-        raise ValueError('Found no telemetry within %d days of %s'
-                         % (days, str(tstart)))
-
-    outnames = ['date'] + [name_map.get(x, x) for x in msids]
-    vals = {name_map.get(x, x): msidset[x].vals for x in msids}
-    vals['date'] = msidset.times
-    out = Ska.Numpy.structured_array(vals, colnames=outnames)
-
-    return out
-
-def plot_two(fig_id, x, y, x2, y2,
-             linestyle='-', linestyle2='-',
-             color='blue', color2='magenta',
-             ylim=None, ylim2=None,
-             xlabel='', ylabel='', ylabel2='', title='',
-             figsize=(7, 3.5),
-             ):
-    """Plot two quantities with a date x-axis"""
-    xt = Ska.Matplotlib.cxctime2plotdate(x)
-    fig = plt.figure(fig_id, figsize=figsize)
-    fig.clf()
-    ax = fig.add_subplot(1, 1, 1)
-    ax.plot_date(xt, y, fmt='-', linestyle=linestyle, color=color)
-    ax.set_xlim(min(xt), max(xt))
-    if ylim:
-        ax.set_ylim(*ylim)
-    ax.set_xlabel(xlabel)
-    ax.set_ylabel(ylabel)
-    ax.set_title(title)
-    ax.grid()
-
-    ax2 = ax.twinx()
-
-    xt2 = Ska.Matplotlib.cxctime2plotdate(x2)
-    ax2.plot_date(xt2, y2, fmt='-', linestyle=linestyle2, color=color2)
-    ax2.set_xlim(min(xt), max(xt))
-    if ylim2:
-        ax2.set_ylim(*ylim2)
-    ax2.set_ylabel(ylabel2, color=color2)
-    ax2.xaxis.set_visible(False)
-
-    Ska.Matplotlib.set_time_ticks(ax)
-    [label.set_rotation(30) for label in ax.xaxis.get_ticklabels()]
-    [label.set_color(color2) for label in ax2.yaxis.get_ticklabels()]
-
-    fig.subplots_adjust(bottom=0.22)
-
-    return {'fig': fig, 'ax': ax, 'ax2': ax2}
